requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.5
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-leaflet==0.27.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
GDAL==3.1.4
importlib-metadata==2.1.1
Pillow==8.1.0
psycopg2==2.8.6
pytz==2020.5
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
whitenoise==5.2.0
zipp==3.4.0
